Test Some Binary Size Optimization Stuff
