+++ Cargo.toml
@@ -7,11 +7,12 @@ edition = "2018"
 members = [
     ".",
     "playground",
+    "wasmtest",
 ]
 
 [dependencies]
 # Data structures
-uuid = { version = "0.8.2", features = ["v4", "serde"] }
+uuid = { version = "0.8.2", features = ["v4", "serde", "wasm-bindgen"] }
 radix_trie = { version = "0.2", features = ["serde"] }
 once_cell = "1.8.0"
 scc = "0.5.2"
@@ -33,3 +34,13 @@ futures-lite = "1.12.0"
 blocking = "1.0.2"
 async-trait = "0.1.51"
 async-timer = "1.0.0-beta.7"
+getrandom = { version = "*", features = ["js"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wee_alloc = "0.4.5"
+
+[profile.release]
+lto = true
+opt-level = "z"
+panic = "abort"
+debug = false
