+++ playground/Cargo.toml
@@ -5,9 +5,9 @@ edition = "2018"
 
 [dependencies]
 rod = { path = "../"}
-serde_json = "1.0.67"
 async-executor = "1.4.1"
 futures-lite = "1.12.0"
+serde_json = "1.0.67"
 anyhow = "1.0.44"
 async-timer = "1.0.0-beta.7"
 tracing-subscriber = "0.2.22"
