+++ src/lib.rs
@@ -3,30 +3,12 @@ pub mod graph;
 pub mod protocol;
 pub mod store;
 
-pub(crate) mod executor {
-    use std::{future::Future, panic, thread};
+pub(crate) mod executor;
 
-    use async_executor::{Executor, Task};
-    use futures_lite::future;
-    use once_cell::sync::Lazy;
+#[cfg(target_arch = "wasm32")]
+extern crate wee_alloc;
 
-    pub fn spawn<T: Send + 'static>(future: impl Future<Output = T> + Send + 'static) -> Task<T> {
-        static GLOBAL: Lazy<Executor<'_>> = Lazy::new(|| {
-            for n in 1..=num_cpus::get() {
-                thread::Builder::new()
-                    .name(format!("rod-worker-{}", n))
-                    .spawn(|| loop {
-                        panic::catch_unwind(|| {
-                            future::block_on(GLOBAL.run(future::pending::<()>()))
-                        })
-                        .ok();
-                    })
-                    .expect("cannot spawn executor thread");
-            }
-
-            Executor::new()
-        });
-
-        GLOBAL.spawn(future)
-    }
-}
+// Use `wee_alloc` as the global allocator for WASM
+#[cfg(target_arch = "wasm32")]
+#[global_allocator]
+static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
