+++ wasmtest/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "wasmtest"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rod = { path = "../"}
+async-executor = "1.4.1"
+futures-lite = "1.12.0"
+anyhow = "1.0.44"
